bare minimul functional movie recommender with flask framework built
