+++ flask_app/templates/recommendation.html
@@ -6,7 +6,7 @@
   </head>
   <body>
 
-    <p>Here is what you should watch:</p>
+    <p>Your recommendations, enjoy!:</p>
     <ul>
       {% for m in movies %}
       <li> {{ m }} </li>
